+++ extensions/tidyread---streamline-your-daily-reading/src/components/DigestDetail.tsx
@@ -1,6 +1,7 @@
 import { Action, Detail } from "@raycast/api";
 import { Digest } from "../types";
 import CustomActionPanel from "./CustomActionPanel";
+import SharableLinkAction from "./SharableLinkAction";
 // import dayjs from "dayjs";
 
 export default function DigestDetail({ digest }: { digest: Digest }) {
@@ -14,6 +15,11 @@ export default function DigestDetail({ digest }: { digest: Digest }) {
       actions={
         <CustomActionPanel>
           <Action.CopyToClipboard title="Copy Content" content={md} />
+          <SharableLinkAction
+            actionTitle="Share This Digest"
+            articleTitle={digest.title}
+            articleContent={digest.content}
+          />
         </CustomActionPanel>
       }
       markdown={md}