+++ extensions/tidyread---streamline-your-daily-reading/src/daily-read.command.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { Source } from "./types";
-import { filterByShownStatus, shell } from "./utils/util";
+import { filterByShownStatus, shell, sleep } from "./utils/util";
 import {
   Action,
   Color,
@@ -13,8 +13,9 @@ import {
   launchCommand,
   showToast,
   useNavigation,
+  LaunchProps,
 } from "@raycast/api";
-import { bizGenDigest, categorizeSources } from "./utils/biz";
+import { addUtmSourceToUrl, bizGenDigest, categorizeSources } from "./utils/biz";
 import { NO_API_KEY, NO_FEEDS, matchError } from "./utils/error";
 import { getSources, getTodaysDigest } from "./store";
 import { capitalize } from "lodash";
@@ -24,17 +25,73 @@ import DigestDetail from "./components/DigestDetail";
 import SharableLinkAction from "./components/SharableLinkAction";
 import CustomActionPanel from "./components/CustomActionPanel";
 
-export default function DailyReadCommand() {
+export default function DailyReadCommand(props: LaunchProps<{ launchContext: { autoGenDigest: boolean } }>) {
+  const autoGenDigest = props?.launchContext?.autoGenDigest ?? false;
   const [todayItems, setTodayItems] = useState<Source[]>();
   const [otherItems, setOtherItems] = useState<Source[]>();
   const itemsLength = (todayItems?.length ?? 0) + (otherItems?.length ?? 0);
-  const { data: todaysDigest = false, revalidate } = usePromise(getTodaysDigest);
+  const { data: todaysDigest, revalidate } = usePromise(getTodaysDigest);
 
   const { push } = useNavigation();
 
   useEffect(() => {
     loadSources();
   }, []);
+
+  const handleGenDigest = async (type: "manual" | "auto" = "manual") => {
+    try {
+      showToast(
+        Toast.Style.Animated,
+        "Generating Digest",
+        "This may take some time, depending on the response speed of the AI API and your setting. And it may fail if you deactivate the Raycast window for an unknown reason.",
+      );
+      await bizGenDigest(type);
+      showToast(Toast.Style.Success, "Generating Success");
+      await revalidate();
+    } catch (err: any) {
+      if (matchError(err, NO_API_KEY)) {
+        showToast(Toast.Style.Failure, "Generating Failed");
+        const markdown = "API key not found. Press Enter to update it in command preferences and try again.";
+
+        push(<Detail markdown={markdown} actions={<CustomActionPanel />} />);
+        return;
+      }
+
+      if (matchError(err, NO_FEEDS)) {
+        showToast(Toast.Style.Failure, "Generating Failed");
+        const markdown =
+          "No RSS link found in today's sources, please add some and try again. Press Enter to manage your sources.";
+
+        push(
+          <Detail markdown={markdown} actions={<CustomActionPanel>{manageSourceListActionNode}</CustomActionPanel>} />,
+        );
+        return;
+      }
+
+      if (matchError(err, "ECONNRESET")) {
+        showToast(Toast.Style.Failure, "Network Error", "Check your network and try again.");
+        return;
+      }
+
+      if (matchError(err, "timed out")) {
+        showToast(Toast.Style.Failure, "Request Timeout", "Check your network, or add http proxy and try again.");
+        return;
+      }
+
+      showToast(Toast.Style.Failure, "Error", err.message);
+    }
+  };
+
+  useEffect(() => {
+    async function fn() {
+      if (autoGenDigest) {
+        await sleep(500);
+        await handleGenDigest("auto");
+      }
+    }
+
+    fn();
+  }, [autoGenDigest]);
 
   const loadSources = async () => {
     const items = await getSources();
@@ -44,56 +101,15 @@ export default function DailyReadCommand() {
   };
 
   const openMultipleUrls = async (items: Source[]) => {
-    const urls = items.map((item) => item.url);
+    const urls = items.map((item) => addUtmSourceToUrl(item.url));
     await shell(`open ${urls.join(" ")}`);
   };
 
   const generateDigestActionNode = (
     <Action
-      title={todayItems ? "Regenerate Today's Digest" : "Generate Today's Digest"}
+      title={todaysDigest ? "Regenerate Today's Digest" : "Generate Today's Digest"}
       icon={Icon.Stars}
-      onAction={async () => {
-        try {
-          showToast(Toast.Style.Animated, "Generating Digest");
-          await bizGenDigest("manual");
-          showToast(Toast.Style.Success, "Generating Success");
-          await revalidate();
-        } catch (err: any) {
-          if (matchError(err, NO_API_KEY)) {
-            showToast(Toast.Style.Failure, "Generating Failed");
-            const markdown = "API key not found. Press Enter to update it in command preferences and try again.";
-
-            push(<Detail markdown={markdown} actions={<CustomActionPanel />} />);
-            return;
-          }
-
-          if (matchError(err, NO_FEEDS)) {
-            showToast(Toast.Style.Failure, "Generating Failed");
-            const markdown =
-              "No RSS link found in today's sources, please add some and try again. Press Enter to manage your sources.";
-
-            push(
-              <Detail
-                markdown={markdown}
-                actions={<CustomActionPanel>{manageSourceListActionNode}</CustomActionPanel>}
-              />,
-            );
-            return;
-          }
-
-          if (matchError(err, "ECONNRESET")) {
-            showToast(Toast.Style.Failure, "Network Error", "Check your network and try again.");
-            return;
-          }
-
-          if (matchError(err, "timed out")) {
-            showToast(Toast.Style.Failure, "Request Timeout", "Check your network, or add http proxy and try again.");
-            return;
-          }
-
-          showToast(Toast.Style.Failure, "Error", err.message);
-        }
-      }}
+      onAction={handleGenDigest}
       shortcut={{ modifiers: ["cmd"], key: "d" }}
     />
   );
@@ -184,7 +200,7 @@ export default function DailyReadCommand() {
                 ])}
                 actions={
                   <CustomActionPanel>
-                    <Action.OpenInBrowser url={item.url} title="Open URL" />
+                    <Action.OpenInBrowser url={addUtmSourceToUrl(item.url)} title="Open URL" />
                     <Action
                       icon={Icon.ArrowNe}
                       title="Open All Today's Sources"
@@ -234,7 +250,7 @@ export default function DailyReadCommand() {
                 ])}
                 actions={
                   <CustomActionPanel>
-                    <Action.OpenInBrowser url={item.url} title="Open URL" />
+                    <Action.OpenInBrowser url={addUtmSourceToUrl(item.url)} title="Open URL" />
                     <Action
                       icon={Icon.ArrowNe}
                       title="Open All Other's Sources"