+++ extensions/tidyread---streamline-your-daily-reading/CHANGELOG.md
@@ -1,3 +1,13 @@
 # TidyRead - Streamline Your Daily Reading Changelog
 
+## [Onboarding Page] - 2024-01-24
+- add quickly generating sources by user interest
+- add retryCount and retryDelay setting to decrease article summarizing error 
+- improve speed of batch import sources
+- improve error handling and error message
+- add Delete All Sources
+- add doc for two commands
+- fix fetch large html content memory leak
+
 ## [Initial Version] - 2024-01-06
+init